chore(clerk-react,shared): Extract and move contexts from clerk-react to shared
