+++ packages/react/src/contexts/ClientContext.tsx
@@ -1,9 +1,1 @@
-import { createContextAndHook } from '@clerk/shared';
-import { ClientResource } from '@clerk/types';
-
-/**
- * @internal
- */
-export const [ClientContext, useClientContext] = createContextAndHook<ClientResource | undefined | null>(
-  'ClientContext',
-);
+export { ClientContext, useClientContext } from '@clerk/shared';
