+++ packages/react/src/contexts/IsomorphicClerkContext.tsx
@@ -1,9 +1,3 @@
-import { createContextAndHook } from '@clerk/shared';
+import { ClerkInstanceContext, useClerkInstanceContext } from '@clerk/shared';
 
-import IsomorphicClerk from '../isomorphicClerk';
-
-/**
- * @internal
- */
-export const [IsomorphicClerkContext, useIsomorphicClerkContext] =
-  createContextAndHook<IsomorphicClerk>('IsomorphicClerkContext');
+export const [IsomorphicClerkContext, useIsomorphicClerkContext] = [ClerkInstanceContext, useClerkInstanceContext];
