+++ packages/react/src/contexts/UserContext.tsx
@@ -1,7 +1,1 @@
-import { createContextAndHook } from '@clerk/shared';
-import { UserResource } from '@clerk/types';
-
-/**
- * @internal
- */
-export const [UserContext, useUserContext] = createContextAndHook<UserResource | null | undefined>('UserContext');
+export { UserContext, useUserContext } from '@clerk/shared';
