+++ packages/shared/package.json
@@ -23,7 +23,8 @@
     "js-cookie": "^3.0.1",
     "tsup": "^6.2.3",
     "typescript": "*",
-    "vitest": "^0.23.4"
+    "vitest": "^0.23.4",
+    "swr": "^1.3.0"
   },
   "peerDependencies": {
     "@testing-library/dom": "^7.28.1",
