+++ packages/shared/src/hooks/index.ts
@@ -1,1 +1,5 @@
 export * from './createContextAndHook';
+export * from './contexts';
+export * from './useOrganization';
+export * from './useOrganizationList';
+export * from './useOrganizations';
